@@ -5,6 +5,8 @@
 plugins {
     id 'java-library'
     id 'maven-publish'
+    id 'org.springframework.boot'
+    id 'io.spring.dependency-management'
 }
 
 repositories {
@@ -15,12 +17,12 @@ repositories {
 }
 
 dependencies {
-    testImplementation 'org.springframework.boot:spring-boot-starter-test:2.7.2'
+    testImplementation 'org.springframework.boot:spring-boot-starter-test:3.1.5'
 }
 
 group = 'github.artbite008'
 version = '1.0.0-SNAPSHOT'
-java.sourceCompatibility = JavaVersion.VERSION_1_8
+java.sourceCompatibility = JavaVersion.VERSION_17
 
 publishing {
     publications {
@@ -32,6 +34,7 @@ publishing {
 
 tasks.withType(JavaCompile) {
     options.encoding = 'UTF-8'
+    options.compilerArgs << '-parameters'
 }
 
 tasks.withType(Javadoc) {
