@@ -1,8 +1,11 @@
 package com.payment.simulator.server.entity;
 
 import lombok.Data;
-import javax.persistence.*;
+import jakarta.persistence.*;
 import java.util.Date;
+
+
+import jakarta.persistence.*;
 
 @Data
 @Entity
