@@ -1,7 +1,7 @@
 package com.payment.simulator.server.entity;
 
 import lombok.Data;
-import javax.persistence.*;
+import jakarta.persistence.*;
 import java.util.Date;
 
 @Data
